+++ src/main/java/com/bizcub/simpleDatapacks/mixin/ReloadCommandMixin.java
@@ -10,26 +10,15 @@ import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
-
-import java.util.Collection;
 
 import net.minecraft.server.commands.ReloadCommand;
-import net.minecraft.server.packs.repository.PackRepository;
-import net.minecraft.world.level.storage.WorldData;
 
 @Mixin(ReloadCommand.class)
 public class ReloadCommandMixin {
 
     @Unique private static boolean simpleDatapacks$shouldSend = false;
 
-    @Inject(method = "discoverNewPacks", at = @At("RETURN"), cancellable = true)
-    private static void preventAutoLoading(PackRepository packRepository, WorldData worldData, Collection<String> enabledDataPacks, CallbackInfoReturnable<Collection<String>> cir) {
-        if (Compat.isClothConfigLoaded() && Configs.getInstance().globalDatapacks) return;
-        cir.setReturnValue(enabledDataPacks);
-    }
-
-    @Inject(method = "reloadPacks", at = @At(value = "HEAD"))
+    @Inject(method = "reloadPacks", at = @At("HEAD"))
     private static void shouldSend(CallbackInfo ci) {
         simpleDatapacks$shouldSend = true;
     }